api implement are done

+++ src/main/java/com/eagle/mailer/entity/sequence/RecipientSequenceEntity.java
@@ -1,5 +1,7 @@
 package com.eagle.mailer.entity.sequence;
 
+import java.math.BigInteger;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -9,5 +11,5 @@ import lombok.Data;
 public class RecipientSequenceEntity {
 	@Id
 	private String id;
-	private long seq;
+	private BigInteger seq;
 }
